+++ jscatter/jscatter.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 
 from matplotlib.colors import to_rgba, Normalize, LogNorm, PowerNorm, LinearSegmentedColormap, ListedColormap
-from typing import Dict, Optional, Union, List, Tuple
+from typing import Dict, Optional, Union, List, Tuple, Literal
 
 from .encodings import Encodings
 from .widget import JupyterScatter, SELECTION_DTYPE
@@ -452,6 +452,9 @@ class Scatter():
         self,
         data: Optional[pd.DataFrame] = None,
         use_index: Optional[Union[bool, Undefined]] = UNDEF,
+        reset_scales: Optional[bool] = False,
+        reset_view: Optional[bool] = False,
+        animate: Optional[bool] = False,
         **kwargs
     ) -> Union[Scatter, dict]:
         """
@@ -460,13 +463,22 @@ class Scatter():
         Parameters
         ----------
         data : pd.DataFrame, optional
-            The data frame that holds the x and y coordinates as well as other
-            possible dimensions that can be used for color, size, or opacity
-            encoding.
+            The new data frame that holds the x and y coordinates as well as
+            other possible dimensions that can be used for color, size, or
+            opacity encoding.
         label_index : bool, optional
             If `True` and if the data frame's index is not an instance of
             `RangeIndex` then the selection and filter methods will reference
             points by the data frame's label index instead of the row index.
+        reset_scales : bool, optional
+            If `True`, all scales (and norms) will be reset to the extend of the
+            the new data.
+        reset_view : bool, optional
+            If `True`, the view will be reset to the origin.
+        animate : bool, optional
+            If `True`, if the number of points remain the same, and if
+            `reset_scales` is `False`, the points will transition smoothly. If
+            `reset_view` is `True`, the view will also transition smoothly.
 
         Returns
         -------
@@ -493,6 +505,8 @@ class Scatter():
         {'data': <pandas.DataFrame>, 'use_label_index': False}
         """
         if data is not None:
+            old_n = self._n
+
             try:
                 self._n = len(data)
                 self._data = data
@@ -503,6 +517,35 @@ class Scatter():
                 return self
 
             self._points = np.zeros((self._n, 6))
+
+            same_n = old_n == self._n
+
+            if reset_scales == True:
+                self._x_scale.vmin = None
+                self._x_scale.vmax = None
+                self._y_scale.vmin = None
+                self._y_scale.vmax = None
+                self._color_norm.vmin = None
+                self._color_norm.vmax = None
+                self._opacity_norm.vmin = None
+                self._opacity_norm.vmax = None
+                self._size_norm.vmin = None
+                self._size_norm.vmax = None
+                self._connection_color_norm.vmin = None
+                self._connection_color_norm.vmax = None
+                self._connection_opacity_norm.vmin = None
+                self._connection_opacity_norm.vmax = None
+                self._connection_size_norm.vmin = None
+                self._connection_size_norm.vmax = None
+
+            if 'skip_widget_update' not in kwargs:
+                # We have to update the following widget props now to ensure
+                # that the update propagated to the widget prior to updating the
+                # points below. This seems to be a bug in Jupyter Lab but I
+                # noticed that the order in which updates are registered on the
+                # front-end does *not* correspond to the call order in Python!
+                self.update_widget('transition_points', animate and same_n)
+                self.update_widget('prevent_filter_reset', False)
 
             self.x(skip_widget_update=True, **self.x())
             self.y(skip_widget_update=True, **self.y())
@@ -515,12 +558,25 @@ class Scatter():
             self.connection_size(skip_widget_update=True, **self.connection_size())
 
             if 'skip_widget_update' not in kwargs:
-                self.update_widget('prevent_filter_reset', False)
                 self.update_widget('points', self.get_point_list())
+                self.update_widget('transition_points', animate == True)
+                self.update_widget('x_domain', self._x_domain)
+                self.update_widget('x_scale_domain', self._x_scale_domain)
+                self.update_widget('y_domain', self._y_domain)
+                self.update_widget('y_scale_domain', self._y_scale_domain)
+
+            if reset_view:
+                if reset_scales:
+                    self.widget.reset_view()
+                else:
+                    animation = 3000 if animate and same_n else 0
+                    self.widget.reset_view(
+                        animation=animation,
+                        data_extent=True,
+                    )
 
         if use_index is not UNDEF:
             self._data_use_index = use_index
-
 
         if any_not([data, use_index], UNDEF):
             return self
@@ -576,18 +632,17 @@ class Scatter():
             if scale is None or scale == 'linear':
                 self._x_scale = create_default_norm()
             elif scale == 'time':
-                self._x_scale = create_default_norm(True)
+                self._x_scale = create_default_norm(is_time=True)
             elif scale == 'log':
-                self._x_scale = LogNorm(clip=True)
+                self._x_scale = LogNorm()
             elif scale == 'pow':
-                self._x_scale = PowerNorm(2, clip=True)
-            elif isinstance(scale, LogNorm) or isinstance(scale, PowerNorm):
+                self._x_scale = PowerNorm(2)
+            elif isinstance(scale, LogNorm) or isinstance(scale, PowerNorm) or isinstance(scale, Normalize):
                 self._x_scale = scale
-                self._x_scale.clip = True
             else:
                 try:
                     vmin, vmax = scale
-                    self._x_scale = Normalize(vmin, vmax, clip=True)
+                    self._x_scale = Normalize(vmin, vmax)
                 except:
                     pass
 
@@ -604,7 +659,8 @@ class Scatter():
                 self._x_by = 'Custom X Data'
 
         if x is not UNDEF or scale is not UNDEF:
-            self.update_widget('prevent_filter_reset', True)
+            if 'skip_widget_update' not in kwargs:
+                self.update_widget('prevent_filter_reset', True)
             self._points[:, 0] = zerofy_missing_values(self.x_data.values, 'X')
 
             self._x_min = np.min(self._points[:, 0])
@@ -692,18 +748,17 @@ class Scatter():
             if scale is None or scale == 'linear':
                 self._y_scale = create_default_norm()
             elif scale == 'time':
-                self._y_scale = create_default_norm(True)
+                self._y_scale = create_default_norm(is_time=True)
             elif scale == 'log':
-                self._y_scale = LogNorm(clip=True)
+                self._y_scale = LogNorm()
             elif scale == 'pow':
-                self._y_scale = PowerNorm(2, clip=True)
-            elif isinstance(scale, LogNorm) or isinstance(scale, PowerNorm):
+                self._y_scale = PowerNorm(2)
+            elif isinstance(scale, LogNorm) or isinstance(scale, PowerNorm) or isinstance(scale, Normalize):
                 self._y_scale = scale
-                self._y_scale.clip = True
             else:
                 try:
                     vmin, vmax = scale
-                    self._y_scale = Normalize(vmin, vmax, clip=True)
+                    self._y_scale = Normalize(vmin, vmax)
                 except:
                     pass
 
@@ -1077,13 +1132,12 @@ class Scatter():
             if callable(norm):
                 try:
                     self._color_norm = norm
-                    self._color_norm.clip = True
                 except:
                     pass
             else:
                 try:
                     vmin, vmax = norm
-                    self._color_norm = Normalize(vmin, vmax, clip=True)
+                    self._color_norm = Normalize(vmin, vmax)
                 except:
                     if norm is None:
                         self._color_norm = create_default_norm()
@@ -1340,13 +1394,12 @@ class Scatter():
             if callable(norm):
                 try:
                     self._opacity_norm = norm
-                    self._opacity_norm.clip = True
                 except:
                     pass
             else:
                 try:
                     vmin, vmax = norm
-                    self._opacity_norm = Normalize(vmin, vmax, clip=True)
+                    self._opacity_norm = Normalize(vmin, vmax)
                 except:
                     if norm is None:
                         # Reset to default value
@@ -1574,13 +1627,12 @@ class Scatter():
             if callable(norm):
                 try:
                     self._size_norm = norm
-                    self._size_norm.clip = True
                 except:
                     pass
             else:
                 try:
                     vmin, vmax = norm
-                    self._size_norm = Normalize(vmin, vmax, clip=True)
+                    self._size_norm = Normalize(vmin, vmax)
                 except:
                     if norm is not None:
                         self._size_norm = create_default_norm()
@@ -1954,13 +2006,12 @@ class Scatter():
             if callable(norm):
                 try:
                     self._connection_color_norm = norm
-                    self._connection_color_norm.clip = True
                 except:
                     pass
             else:
                 try:
                     vmin, vmax = norm
-                    self._connection_color_norm = Normalize(vmin, vmax, clip=True)
+                    self._connection_color_norm = Normalize(vmin, vmax)
                 except:
                     if norm is None:
                         self._connection_color_norm = create_default_norm()
@@ -2206,13 +2257,12 @@ class Scatter():
             if callable(norm):
                 try:
                     self._connection_opacity_norm.clip = norm
-                    self._connection_opacity_norm.clip = True
                 except:
                     pass
             else:
                 try:
                     vmin, vmax = norm
-                    self._connection_opacity_norm = Normalize(vmin, vmax, clip=True)
+                    self._connection_opacity_norm = Normalize(vmin, vmax)
                 except:
                     if norm is None:
                         self._connection_opacity_norm = create_default_norm()
@@ -2437,13 +2487,12 @@ class Scatter():
             if callable(norm):
                 try:
                     self._connection_size_norm = norm
-                    self._connection_size_norm.clip = True
                 except:
                     pass
             else:
                 try:
                     vmin, vmax = norm
-                    self._connection_size_norm = Normalize(vmin, vmax, clip=True)
+                    self._connection_size_norm = Normalize(vmin, vmax)
                 except:
                     if norm is None:
                         self._connection_size_norm = create_default_norm()